requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0,<3.0.0
python-multipart>=0.0.7
SQLAlchemy>=2.0.0
pillow>=10.0.0
numpy>=1.20.0
rembg>=2.0.0
ffmpeg-python>=0.2.0
opencv-python-headless>=4.0.0
aiohttp>=3.8.0
aiofiles>=22.0.0
requests>=2.25.0
gunicorn>=21.0.0
python-jose[cryptography]>=3.0.0
passlib[bcrypt]>=1.7.0

# Diffusers stack - letting pip solve dependencies
transformers
diffusers
accelerate
safetensors
torch
torchvision
huggingface_hub
tokenizers

# SMPL Avatar integration
scipy>=1.0.0
cython>=0.29.0
pyrender>=0.1.45
trimesh>=3.0.0
pycocotools>=2.0.0
smplx>=0.1.0

# Integration API dependencies
websocket-client>=1.0.0
websockets>=10.0.0
tqdm>=4.0.0
psutil>=5.0.0
einops>=0.6.0
moviepy>=1.0.0
omegaconf>=2.0.0
